+++ src/lib/downloadCsv.ts
@@ -1,0 +1,15 @@
+
+
+export function downloadFile(blob: Blob, filename: string): void {
+  const url = window.URL.createObjectURL(blob);
+
+  const link = document.createElement("a");
+  link.href = url;
+  link.download = filename;
+
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+
+  window.URL.revokeObjectURL(url);
+}
